edtr-io: enable MathJax for tables
Close #541
