+++ packages/private/mathjax/package.json
@@ -12,6 +12,7 @@
     "tslib": "^2.0.0"
   },
   "devDependencies": {
+    "@types/jquery": "^3.0.0",
     "ts-node": "^9.0.0",
     "tsdx": "^0.14.0",
     "typescript": "^4.0.0"
