+++ packages/private/mathjax/src/index.ts
@@ -19,6 +19,8 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/serlo.org for the canonical source repository
  */
+let mathJaxInitialized = false
+
 export function reprocess(context?: unknown) {
   command('Reprocess', context)
 }
@@ -39,4 +41,59 @@ function command(command: string, context?: unknown) {
   if (MathJax) {
     MathJax.Hub.Queue([command, MathJax.Hub, ...(context ? [context] : [])])
   }
+}
+
+export function initMathJax() {
+  if (typeof window === undefined || mathJaxInitialized) return
+  mathJaxInitialized = true
+  const { $ } = (window as unknown) as { $: JQueryStatic }
+
+  if ($('mathjax-config').length > 0) return
+  const lang = $('html').attr('lang') || 'en'
+  $('head').append(
+    `
+      <script type="text/x-mathjax-config" id="mathjax-config">
+        MathJax.Hub.Config({
+            displayAlign: 'left',
+            extensions: ['tex2jax.js', 'fast-preview.js'],
+            jax: ['input/TeX', 'output/SVG', 'output/CommonHTML'],
+            skipStartupTypeset: true,
+            tex2jax: {
+                inlineMath: [
+                    ['%%', '%%']
+                ],
+                displayMath: [
+                    ['$$', '$$']
+                ],
+            },
+            'HTML-CSS': {
+                scale: 100,
+                linebreaks: {
+                    automatic: true
+                },
+                preferredFont: 'STIX'
+            },
+            SVG: {
+                linebreaks: {
+                    automatic: true
+                },
+                font: 'STIX-Web'
+            },
+            showProcessingMessages: false,
+            TeX: {
+                extensions: ['mediawiki-texvc.js', 'mhchem.js']
+            }
+        });
+      </script>
+    `
+  )
+  $.ajax({
+    url: `https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.9/MathJax.js?config=TeX-AMS-MML_HTMLorMML&locale=${lang}`,
+    dataType: 'script',
+    success() {
+      // Initial typeset
+      typeset()
+    },
+    async: true,
+  })
 }
