+++ packages/public/client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@serlo/client",
-  "version": "17.0.4",
+  "version": "17.0.5",
   "private": true,
   "scripts": {
     "prebuild": "rimraf dist",
