+++ packages/public/editor-renderer/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@serlo/editor-renderer",
-  "version": "11.0.1",
+  "version": "11.0.2",
   "private": true,
   "main": "dist/index.js",
   "scripts": {
